WIP: Refactor and various improvements
