+++ src/widgets/ScrollBar.ts
