+++ tests/unit/all.ts
@@ -1,9 +1,6 @@
-import './Editor';
-import './IconCss';
 import './main';
 import './project';
 import './routing';
-import './Runner';
-import './TreePane';
 import './Workbench';
 import './support/all';
+import './widgets/all';
