+++ tests/unit/widgets/IconCss.ts
@@ -3,8 +3,8 @@ import * as assert from 'intern/chai!assert';
 import harness from '@dojo/test-extras/harness';
 import { v } from '@dojo/widget-core/d';
 import { HNode } from '@dojo/widget-core/interfaces';
-import IconCss, { IconCssProperties } from '../../src/IconCss';
-import icons from '../support/iconJson';
+import IconCss, { IconCssProperties } from '../../../src/widgets/IconCss';
+import icons from '../../support/iconJson';
 
 registerSuite({
 	name: 'IconCss',
