+++ tests/unit/widgets/all.ts
@@ -1,0 +1,5 @@
+import './Editor';
+import './IconCss';
+import './Runner';
+import './ScrollBar';
+import './TreePane';
